requirements.txt
fastapi
uvicorn
apscheduler
loguru
watchdog
pytz
python-dotenv
httpx
Pillow
